requirements.txt
openai==1.10.0
numpy==1.26.4
python-dotenv==1.0.1
fastapi==0.112.0
uvicorn==0.30.6
Jinja2==3.1.4
python-multipart==0.0.9
